LayoutTests/platform/ios-sim/http/tests/misc/webtiming-one-redirect-expected.txt

This test checks that Web Timing reports redirectCount correctly and has reasonable values for redirectStart and redirectCount.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL timing.redirectStart should be >= timing.navigationStart. Was undefined (of type undefined).
FAIL timing.redirectEnd should be >= timing.redirectStart. Was undefined (of type undefined).
FAIL navigation.redirectCount should be 1 (of type number). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
